requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.7.1
frozenlist==1.8.0
git-filter-repo==2.47.0
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dotenv==1.2.2
ruff==0.15.7
SQLAlchemy==2.0.48
SQLAlchemy-Utils==0.42.1
typing_extensions==4.15.0
yarl==1.23.0
